refactor: make classes private
