+++ deepset_cloud_sdk/_api/files.py
@@ -14,7 +14,7 @@ from uuid import UUID
 import structlog
 from httpx import codes
 
-from deepset_cloud_sdk.api.deepset_cloud_api import DeepsetCloudAPI
+from deepset_cloud_sdk._api.deepset_cloud_api import DeepsetCloudAPI
 
 logger = structlog.get_logger(__name__)
