+++ deepset_cloud_sdk/_service/files_service.py
@@ -13,16 +13,16 @@ import structlog
 from tqdm import tqdm
 from yaspin import yaspin
 
-from deepset_cloud_sdk.api.config import CommonConfig
-from deepset_cloud_sdk.api.deepset_cloud_api import DeepsetCloudAPI
-from deepset_cloud_sdk.api.files import File, FilesAPI
-from deepset_cloud_sdk.api.upload_sessions import (
+from deepset_cloud_sdk._api.config import CommonConfig
+from deepset_cloud_sdk._api.deepset_cloud_api import DeepsetCloudAPI
+from deepset_cloud_sdk._api.files import File, FilesAPI
+from deepset_cloud_sdk._api.upload_sessions import (
     UploadSession,
     UploadSessionsAPI,
     WriteMode,
 )
+from deepset_cloud_sdk._s3.upload import S3
 from deepset_cloud_sdk.models import DeepsetCloudFile
-from deepset_cloud_sdk.s3.upload import S3
 
 logger = structlog.get_logger(__name__)
