+++ deepset_cloud_sdk/cli.py
@@ -6,7 +6,7 @@ import typer
 from tabulate import tabulate
 
 from deepset_cloud_sdk.__about__ import __version__
-from deepset_cloud_sdk.api.config import DEFAULT_WORKSPACE_NAME, ENV_FILE_PATH
+from deepset_cloud_sdk._api.config import DEFAULT_WORKSPACE_NAME, ENV_FILE_PATH
 from deepset_cloud_sdk.workflows.sync_client.files import list_files as sync_list_files
 from deepset_cloud_sdk.workflows.sync_client.files import upload
