+++ tests/integration/api/test_integration_files.py
@@ -1,12 +1,9 @@
-from typing import List
-from unittest.mock import Mock
-
 import httpx
 import pytest
 
-from deepset_cloud_sdk.api.config import CommonConfig
-from deepset_cloud_sdk.api.deepset_cloud_api import DeepsetCloudAPI
-from deepset_cloud_sdk.api.files import File, FilesAPI
+from deepset_cloud_sdk._api.config import CommonConfig
+from deepset_cloud_sdk._api.deepset_cloud_api import DeepsetCloudAPI
+from deepset_cloud_sdk._api.files import File, FilesAPI
 
 
 @pytest.fixture
