+++ tests/unit/workflows/async_client/test_async_workflow_files.py
@@ -8,10 +8,10 @@ import pytest
 from _pytest.monkeypatch import MonkeyPatch
 from sniffio import AsyncLibraryNotFoundError
 
-from deepset_cloud_sdk.api.config import DEFAULT_WORKSPACE_NAME
-from deepset_cloud_sdk.api.files import File
-from deepset_cloud_sdk.api.upload_sessions import WriteMode
-from deepset_cloud_sdk.service.files_service import DeepsetCloudFile, FilesService
+from deepset_cloud_sdk._api.config import DEFAULT_WORKSPACE_NAME
+from deepset_cloud_sdk._api.files import File
+from deepset_cloud_sdk._api.upload_sessions import WriteMode
+from deepset_cloud_sdk._service.files_service import DeepsetCloudFile, FilesService
 from deepset_cloud_sdk.workflows.async_client.files import (
     list_files,
     upload,
